Extract report-writer helpers to shrink reporting-screen root
reporting-screen.tsx carried ~590 lines of pure report-writer request/formula/
preview builders (no JSX, no component state) that supplemented rather than left
the root. Move the 32-function leaf cluster (grid/access/filter/metric/formula
request builders, the formula field parser, dataset-preview value generators)
into reporting-screen.report-writer-helpers.ts. The cluster is downward-closed
(no back-references into the root), so there are no circular imports; only 6 of
the 32 are used by the root and re-imported. Root drops from 4021 to 3439 lines.
